Allow editing target for Env infractions on targets reported by the unit + various fixes
