+++ frontend/src/pam/mission/actions/action-control-env.tsx
@@ -10,242 +10,218 @@ import { Action } from '../../../types/action-types'
 import { ControlType } from '../../../types/control-types'
 import { actionTargetTypeLabels, EnvActionControl, vehicleTypeLabels } from '../../../types/env-mission-types'
 import { useParams } from 'react-router-dom'
-import EnvInfractionNewTarget from '../infractions/env-infraction-new-target'
-import EnvInfractionExistingTarget from '../infractions/env-infraction-existing-target'
+import EnvInfractionAddNewTarget from '../infractions/env-infraction-add-new-target.tsx'
+import EnvInfractionExistingTargets from '../infractions/env-infraction-existing-targets.tsx'
 import useActionById from "./use-action-by-id.tsx";
 import { extractLatLonFromMultiPoint } from "../../../utils/geometry.ts";
 
 interface ActionControlPropsEnv {
-  action: Action
+    action: Action
 }
 
 const ActionControlEnv: React.FC<ActionControlPropsEnv> = ({action}) => {
-  const {missionId, actionId} = useParams()
+    const {missionId, actionId} = useParams()
 
-  const {
-    data: envAction,
-    loading,
-    error,
-  } = useActionById(actionId, missionId, action.source, action.type)
+    const {
+        data: envAction,
+        loading,
+        error,
+    } = useActionById(actionId, missionId, action.source, action.type)
 
-  if (loading) {
-    return (
-      <div>loading</div>
-    )
-  }
-  if (error) {
-    return (
-      <div>error</div>
-    )
-  }
-  if (envAction) {
-    const actionData: EnvActionControl = envAction.data
-    return (
-      <Stack direction="column" spacing={'2rem'} alignItems="flex-start" style={{width: '100%'}}>
-        <Stack.Item style={{width: '100%'}}>
-          <Stack direction="row" spacing={'0.5rem'} style={{width: '100%'}}>
-            <Stack.Item alignSelf="baseline">
-              <Icon.ControlUnit color={THEME.color.charcoal} size={20}/>
-            </Stack.Item>
-            <Stack.Item grow={2}>
-              <Text as="h2" weight="bold">
-                Contrôles {envAction.startDateTimeUtc && `(${formatDateTimeForFrenchHumans(envAction.startDateTimeUtc)})`}
-              </Text>
-            </Stack.Item>
-          </Stack>
-        </Stack.Item>
-        <Stack.Item style={{width: '100%'}}>
-          <Label>Thématique de contrôle</Label>
-          <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
-            {actionData?.themes[0].theme ?? 'inconnue'}
-          </Text>
-        </Stack.Item>
-        <Stack.Item style={{width: '100%'}}>
-          <Label>Sous-thématiques de contrôle</Label>
-          <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
-            {actionData?.themes[0].subThemes?.length
-              ? actionData?.themes[0].subThemes?.join(', ')
-              : 'inconnues'}
-          </Text>
-        </Stack.Item>
-        <Stack.Item style={{width: '100%'}}>
-          <Label>Date et heure</Label>
-          <DatePicker
-            defaultValue={envAction.startDateTimeUtc}
-            // label="Date et heure"
-            withTime={true}
-            isCompact={false}
-            isLight={true}
-            name="startDateTimeUtc"
-            readOnly={true}
-            disabled={true}
-          />
-        </Stack.Item>
-        <Stack.Item>
-          <Label>Lieu du contrôle</Label>
-          {true && (() => {
-            const [latitude, longitude] = extractLatLonFromMultiPoint(actionData?.geom);
-            return (
-              <CoordinatesInput
-                defaultValue={[
-                  latitude as any,
-                  longitude as any
-                ]}
-                coordinatesFormat={CoordinatesFormat.DECIMAL_DEGREES}
-                // label="Lieu du contrôle"
-                // isLight={true}
-                disabled={true}
-              />
-            );
-          })()}
-
-
-          {/*<MultiZoneEditor*/}
-          {/*    addButtonLabel="Add a zone"*/}
-          {/*    disabled*/}
-          {/*    error=""*/}
-          {/*    initialZone={{*/}
-          {/*        name: 'Polygone dessiné'*/}
-          {/*    }}*/}
-          {/*    isLight={true}*/}
-          {/*    label="Some zones"*/}
-          {/*    labelPropName="name"*/}
-          {/*    defaultValue={actionData?.geom}*/}
-          {/*    onAdd={function noRefCheck() {*/}
-          {/*    }}*/}
-          {/*    onCenter={function noRefCheck() {*/}
-          {/*    }}*/}
-          {/*    onChange={function noRefCheck() {*/}
-          {/*    }}*/}
-          {/*    onDelete={function noRefCheck() {*/}
-          {/*    }}*/}
-          {/*    onEdit={function noRefCheck() {*/}
-          {/*    }}*/}
-          {/*/>*/}
-        </Stack.Item>
-        <Stack.Item style={{width: '100%'}}>
-          <Divider style={{backgroundColor: THEME.color.charcoal}}/>
-        </Stack.Item>
-        <Stack.Item style={{width: '100%'}}>
-          <Stack direction="column" alignItems="flex-start" spacing={'2rem'} style={{width: '100%'}}>
-            <Stack.Item style={{width: '100%'}}>
-              <Stack direction="row" alignItems="flex-start" spacing={'2rem'} style={{width: '100%'}}>
-                <Stack.Item style={{width: '33%'}}>
-                  <Stack direction="column" spacing={'1rem'} alignItems="flex-start">
-                    <Stack.Item>
-                      <Label>Nbre total de contrôles</Label>
-                      <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
-                        {actionData?.actionNumberOfControls ?? 'inconnu'}
-                      </Text>
-                    </Stack.Item>
-                    <Stack.Item>
-                      <Label>Type de cible</Label>
-                      <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
-                        {!!actionData?.actionTargetType
-                          ? actionTargetTypeLabels[actionData?.actionTargetType].libelle
-                          : 'inconnu'}
-                      </Text>
-                    </Stack.Item>
-                    <Stack.Item>
-                      <Label>Type de véhicule</Label>
-                      <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
-                        {!!actionData?.vehicleType ? vehicleTypeLabels[actionData?.vehicleType].libelle : 'inconnu'}
-                      </Text>
-                    </Stack.Item>
-                    <Stack.Item>
-                      <Label>Observations</Label>
-                      <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
-                        {actionData?.observations ?? 'aucunes'}
-                      </Text>
-                    </Stack.Item>
-                  </Stack>
+    if (loading) {
+        return (
+            <div>loading</div>
+        )
+    }
+    if (error) {
+        return (
+            <div>error</div>
+        )
+    }
+    if (envAction) {
+        const actionData: EnvActionControl = envAction.data
+        return (
+            <Stack direction="column" spacing={'2rem'} alignItems="flex-start" style={{width: '100%'}}>
+                <Stack.Item style={{width: '100%'}}>
+                    <Stack direction="row" spacing={'0.5rem'} style={{width: '100%'}}>
+                        <Stack.Item alignSelf="baseline">
+                            <Icon.ControlUnit color={THEME.color.charcoal} size={20}/>
+                        </Stack.Item>
+                        <Stack.Item grow={2}>
+                            <Text as="h2" weight="bold">
+                                Contrôles {envAction.startDateTimeUtc && `(${formatDateTimeForFrenchHumans(envAction.startDateTimeUtc)})`}
+                            </Text>
+                        </Stack.Item>
+                    </Stack>
                 </Stack.Item>
-
-                <Stack.Item style={{width: '67%'}}>
-                  <Stack direction="column" alignItems="flex-start" spacing={'1.5rem'}
-                         style={{width: '100%'}}>
-                    {(actionData?.controlsToComplete?.length || 0) > 0 && (
-                      <Stack.Item alignSelf="flex-end">
-                        <ControlsToCompleteTag
-                          amountOfControlsToComplete={
-                            actionData?.controlsToComplete?.length
-                          }
-                          isLight={true}
-                        />
-                      </Stack.Item>
-                    )}
-
-                    <Stack.Item>
-                      <Text as="h3" weight="bold" color={THEME.color.gunMetal}>
-                        dont...
-                      </Text>
-                    </Stack.Item>
-                    <Stack.Item style={{width: '100%'}}>
-                      <EnvControlForm
-                        controlType={ControlType.ADMINISTRATIVE}
-                        data={actionData?.controlAdministrative}
-                        maxAmountOfControls={actionData?.actionNumberOfControls}
-                        shouldCompleteControl={
-                          !!actionData?.controlsToComplete?.includes(
-                            ControlType.ADMINISTRATIVE
-                          )
-                        }
-                      />
-                    </Stack.Item>
-                    <Stack.Item style={{width: '100%'}}>
-                      <EnvControlForm
-                        controlType={ControlType.NAVIGATION}
-                        data={actionData?.controlNavigation}
-                        maxAmountOfControls={actionData?.actionNumberOfControls}
-                        shouldCompleteControl={
-                          !!actionData?.controlsToComplete?.includes(ControlType.NAVIGATION)
-                        }
-                      />
-                    </Stack.Item>
-                    <Stack.Item style={{width: '100%'}}>
-                      <EnvControlForm
-                        controlType={ControlType.GENS_DE_MER}
-                        data={actionData?.controlGensDeMer}
-                        maxAmountOfControls={actionData?.actionNumberOfControls}
-                        shouldCompleteControl={
-                          !!actionData?.controlsToComplete?.includes(
-                            ControlType.GENS_DE_MER
-                          )
-                        }
-                      />
-                    </Stack.Item>
-                    <Stack.Item style={{width: '100%'}}>
-                      <EnvControlForm
-                        controlType={ControlType.SECURITY}
-                        data={actionData?.controlSecurity}
-                        maxAmountOfControls={actionData?.actionNumberOfControls}
-                        shouldCompleteControl={
-                          !!actionData?.controlsToComplete?.includes(ControlType.SECURITY)
-                        }
-                      />
-                    </Stack.Item>
-                  </Stack>
+                <Stack.Item style={{width: '100%'}}>
+                    <Label>Thématique de contrôle</Label>
+                    <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
+                        {actionData?.themes[0].theme ?? 'inconnue'}
+                    </Text>
                 </Stack.Item>
-              </Stack>
-            </Stack.Item>
-            <Stack.Item style={{width: '100%'}}>
-              <EnvInfractionNewTarget
-                availableControlTypesForInfraction={actionData?.availableControlTypesForInfraction}
-              />
-            </Stack.Item>
-            <Stack.Item style={{width: '100%'}}>
-              <EnvInfractionExistingTarget
-                infractionsByTarget={actionData?.infractions}
-                availableControlTypesForInfraction={actionData?.availableControlTypesForInfraction}
-              />
-            </Stack.Item>
-          </Stack>
-        </Stack.Item>
-      </Stack>
-    )
-  }
+                <Stack.Item style={{width: '100%'}}>
+                    <Label>Sous-thématiques de contrôle</Label>
+                    <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
+                        {actionData?.themes[0].subThemes?.length
+                            ? actionData?.themes[0].subThemes?.join(', ')
+                            : 'inconnues'}
+                    </Text>
+                </Stack.Item>
+                <Stack.Item style={{width: '100%'}}>
+                    <Label>Date et heure</Label>
+                    <DatePicker
+                        defaultValue={envAction.startDateTimeUtc}
+                        // label="Date et heure"
+                        withTime={true}
+                        isCompact={false}
+                        isLight={true}
+                        name="startDateTimeUtc"
+                        readOnly={true}
+                        disabled={true}
+                    />
+                </Stack.Item>
+                <Stack.Item>
+                    <Label>Lieu du contrôle</Label>
+                    {true && (() => {
+                        const [latitude, longitude] = extractLatLonFromMultiPoint(actionData?.geom);
+                        return (
+                            <CoordinatesInput
+                                defaultValue={[
+                                    latitude as any,
+                                    longitude as any
+                                ]}
+                                coordinatesFormat={CoordinatesFormat.DECIMAL_DEGREES}
+                                // label="Lieu du contrôle"
+                                // isLight={true}
+                                disabled={true}
+                            />
+                        );
+                    })()}
+                </Stack.Item>
+                <Stack.Item style={{width: '100%'}}>
+                    <Divider style={{backgroundColor: THEME.color.charcoal}}/>
+                </Stack.Item>
+                <Stack.Item style={{width: '100%'}}>
+                    <Stack direction="column" alignItems="flex-start" spacing={'2rem'} style={{width: '100%'}}>
+                        <Stack.Item style={{width: '100%'}}>
+                            <Stack direction="row" alignItems="flex-start" spacing={'2rem'} style={{width: '100%'}}>
+                                <Stack.Item style={{width: '33%'}}>
+                                    <Stack direction="column" spacing={'1rem'} alignItems="flex-start">
+                                        <Stack.Item>
+                                            <Label>Nbre total de contrôles</Label>
+                                            <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
+                                                {actionData?.actionNumberOfControls ?? 'inconnu'}
+                                            </Text>
+                                        </Stack.Item>
+                                        <Stack.Item>
+                                            <Label>Type de cible</Label>
+                                            <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
+                                                {!!actionData?.actionTargetType
+                                                    ? actionTargetTypeLabels[actionData?.actionTargetType].libelle
+                                                    : 'inconnu'}
+                                            </Text>
+                                        </Stack.Item>
+                                        <Stack.Item>
+                                            <Label>Type de véhicule</Label>
+                                            <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
+                                                {!!actionData?.vehicleType ? vehicleTypeLabels[actionData?.vehicleType].libelle : 'inconnu'}
+                                            </Text>
+                                        </Stack.Item>
+                                        <Stack.Item>
+                                            <Label>Observations</Label>
+                                            <Text as="h3" weight="medium" color={THEME.color.gunMetal}>
+                                                {actionData?.observations ?? 'aucunes'}
+                                            </Text>
+                                        </Stack.Item>
+                                    </Stack>
+                                </Stack.Item>
 
-  return null
+                                <Stack.Item style={{width: '67%'}}>
+                                    <Stack direction="column" alignItems="flex-start" spacing={'1.5rem'}
+                                           style={{width: '100%'}}>
+                                        {(actionData?.controlsToComplete?.length || 0) > 0 && (
+                                            <Stack.Item alignSelf="flex-end">
+                                                <ControlsToCompleteTag
+                                                    amountOfControlsToComplete={
+                                                        actionData?.controlsToComplete?.length
+                                                    }
+                                                    isLight={true}
+                                                />
+                                            </Stack.Item>
+                                        )}
+
+                                        <Stack.Item>
+                                            <Text as="h3" weight="bold" color={THEME.color.gunMetal}>
+                                                dont...
+                                            </Text>
+                                        </Stack.Item>
+                                        <Stack.Item style={{width: '100%'}}>
+                                            <EnvControlForm
+                                                controlType={ControlType.ADMINISTRATIVE}
+                                                data={actionData?.controlAdministrative}
+                                                maxAmountOfControls={actionData?.actionNumberOfControls}
+                                                shouldCompleteControl={
+                                                    !!actionData?.controlsToComplete?.includes(
+                                                        ControlType.ADMINISTRATIVE
+                                                    )
+                                                }
+                                            />
+                                        </Stack.Item>
+                                        <Stack.Item style={{width: '100%'}}>
+                                            <EnvControlForm
+                                                controlType={ControlType.NAVIGATION}
+                                                data={actionData?.controlNavigation}
+                                                maxAmountOfControls={actionData?.actionNumberOfControls}
+                                                shouldCompleteControl={
+                                                    !!actionData?.controlsToComplete?.includes(ControlType.NAVIGATION)
+                                                }
+                                            />
+                                        </Stack.Item>
+                                        <Stack.Item style={{width: '100%'}}>
+                                            <EnvControlForm
+                                                controlType={ControlType.GENS_DE_MER}
+                                                data={actionData?.controlGensDeMer}
+                                                maxAmountOfControls={actionData?.actionNumberOfControls}
+                                                shouldCompleteControl={
+                                                    !!actionData?.controlsToComplete?.includes(
+                                                        ControlType.GENS_DE_MER
+                                                    )
+                                                }
+                                            />
+                                        </Stack.Item>
+                                        <Stack.Item style={{width: '100%'}}>
+                                            <EnvControlForm
+                                                controlType={ControlType.SECURITY}
+                                                data={actionData?.controlSecurity}
+                                                maxAmountOfControls={actionData?.actionNumberOfControls}
+                                                shouldCompleteControl={
+                                                    !!actionData?.controlsToComplete?.includes(ControlType.SECURITY)
+                                                }
+                                            />
+                                        </Stack.Item>
+                                    </Stack>
+                                </Stack.Item>
+                            </Stack>
+                        </Stack.Item>
+                        <Stack.Item style={{width: '100%'}}>
+                            <EnvInfractionAddNewTarget
+                                availableControlTypesForInfraction={actionData?.availableControlTypesForInfraction}
+                            />
+                        </Stack.Item>
+                        <Stack.Item style={{width: '100%'}}>
+                            <EnvInfractionExistingTargets
+                                infractionsByTarget={actionData?.infractions}
+                                availableControlTypesForInfraction={actionData?.availableControlTypesForInfraction}
+                            />
+                        </Stack.Item>
+                    </Stack>
+                </Stack.Item>
+            </Stack>
+        )
+    }
+
+    return null
 }
 
 export default ActionControlEnv
