+++ frontend/src/pam/mission/actions/use-action-by-id.tsx
@@ -154,7 +154,7 @@ export const GET_ACTION_BY_ID = gql`
           infractions {
             vesselIdentifier
             vesselType
-            targetAddedInRapportNav
+            targetAddedByUnit
             controlTypesWithInfraction
             infractions {
               id
