+++ frontend/src/pam/mission/infractions/utils.ts
@@ -1,30 +1,32 @@
 import { ControlType } from '../../../types/control-types'
+import { Infraction, InfractionByTarget, InfractionEnvNewTarget } from "../../../types/infraction-types.ts";
+import { VesselSizeEnum, VesselTypeEnum } from "../../../types/env-mission-types.ts";
 
 export const infractionTitleForControlType = (controlType: ControlType): string => {
-  switch (controlType) {
-    case ControlType.ADMINISTRATIVE:
-      return 'Infraction administrative'
-    case ControlType.NAVIGATION:
-      return 'Infraction règles de navigation'
-    case ControlType.SECURITY:
-      return 'Infraction equipements et respect des normes de sécurité'
-    case ControlType.GENS_DE_MER:
-      return 'Infraction administrative gens de mer'
-    default:
-      return ''
-  }
+    switch (controlType) {
+        case ControlType.ADMINISTRATIVE:
+            return 'Infraction administrative'
+        case ControlType.NAVIGATION:
+            return 'Infraction règles de navigation'
+        case ControlType.SECURITY:
+            return 'Infraction equipements et respect des normes de sécurité'
+        case ControlType.GENS_DE_MER:
+            return 'Infraction administrative gens de mer'
+        default:
+            return ''
+    }
 }
 export const infractionButtonTitle = (controlType: ControlType): string => {
-  switch (controlType) {
-    case ControlType.ADMINISTRATIVE:
-      return 'Ajouter une infraction administrative'
-    case ControlType.NAVIGATION:
-      return 'Ajouter une infraction règle de navigation'
-    case ControlType.SECURITY:
-      return 'Ajouter une infraction sécurité'
-    case ControlType.GENS_DE_MER:
-      return 'Ajouter une infraction administrative'
-    default:
-      return ''
-  }
+    switch (controlType) {
+        case ControlType.ADMINISTRATIVE:
+            return 'Ajouter une infraction administrative'
+        case ControlType.NAVIGATION:
+            return 'Ajouter une infraction règle de navigation'
+        case ControlType.SECURITY:
+            return 'Ajouter une infraction sécurité'
+        case ControlType.GENS_DE_MER:
+            return 'Ajouter une infraction administrative'
+        default:
+            return ''
+    }
 }
